requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic>=2.7.4,<3.0.0
pydantic-settings==2.1.0

# HTTP Clients
httpx[http2]>=0.27.0
aiohttp>=3.10.0,<4.0.0
requests>=2.31.0
python-socketio>=5.11.0  # Photon Socket.IO listener
python-engineio>=4.9.0,<5.0.0

# Database / Storage
supabase>=2.9.0
psycopg-binary>=3.2.0

# AI/LLM
openai>=1.54.0
tiktoken>=0.7.0
tenacity==8.2.3
# zep-python removed - using custom HTTP client (zep_client_simple.py) for V3 SDK compatibility
langgraph==0.2.53
langchain-openai==0.2.14
langchain-core>=0.3.43,<1.0.0
composio>=0.9.0,<1.0.0

# Messaging / Reactions
numpy>=1.24.0  # Matching utilities
python-dateutil>=2.8.2
pytz>=2024.1

# Caching / Rate limiting
redis>=4.5.2,<5.0.0
slowapi==0.1.9
cachetools>=5.3.0

# Kafka
aiokafka>=0.10.0,<0.11.0
aws-msk-iam-sasl-signer-python>=1.0.2,<2.0.0

# Monitoring & Logging
sentry-sdk[fastapi]==1.40.0
python-json-logger==2.0.7
psutil==5.9.8

# Testing / Dev
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
black==24.1.1
flake8==7.0.0
mypy==1.8.0

# Payments
stripe>=5.0.0,<14.0.0
